rabbit: Report base and overridden product info separately
Now, `rabbit:product_info/0` will report the base name & version
("RabbitMQ" and its actual version) separately from the final product
name & version (if any).

If the product name & version were not overridden in the configuration or
the environment, they will be missing from the product info return
value.

It allows caller to distinguish situations where the product name &
version were set or not.

An example use case is `rabbitmqctl status`: we want to report the
product name & version only if they were overridden. If they were not,
the "Product name" and "Product version" won't be added to the output.
Only "RabbitMQ version" will be displayed.

To help with the above, the map returned by `rabbit:product_info/0` also
contains a flag, `product_overridden`, set to true if they were.

Finally, the RabbitMQ version is often associated with the Erlang/OTP
release. So include this into the returned map. This will save an
additional RPC call if the remote caller is interested in both.